requirements.txt
replicate
Pillow
requests
python-dotenv
